backend/requirements.txt
flask
mysql-connector-python
flask-cors
paho-mqtt>=2.0.0